Updated CategoryPolicy

+++ app/Http/Controllers/CategoryTaskController.php
@@ -115,8 +115,21 @@ class CategoryTaskController extends Controller
 
     public function destroy(Category $list, Task $task)
     {
+        $this->authorize('delete', $task);
+
         $task->delete();
 
         return redirect()->route('lists.show', [$list->id])->with('success', 'Task Deleted.');
     }
 }
+
+
+
+// public function destroy(Category $list)
+//     {
+//         $this->authorize('delete', $list);
+
+//         $list->delete();
+
+//         return back()->with('success', 'List deleted.');
+//     }
